Updated config, and how script reads from it

+++ WindowsMaintenenceStep2.ps1
@@ -8,26 +8,51 @@ function Write-Completed {
     Write-Host "Completed"
 }
 
-# DISM RestoreHealth
-Write-Header -header "Checking Windows image..."
-DISM /Online /Cleanup-Image /RestoreHealth
+# Import config file for input variables for script
+Write-Header -header "Importing configuration..."
+$config = Get-Content -Raw -Path config.json | ConvertFrom-Json
 Write-Completed
 
-# WinSxS cleanup ResetBase
-Write-Header -header "Cleaning up WinSxS..."
-DISM /Online /Cleanup-Image /StartComponentCleanup /ResetBase
-Write-Completed
+# Only run if previous script has run
+if ($config == 1)
+{
+    # DISM RestoreHealth
+    Write-Header -header "Checking Windows image..."
+    DISM /Online /Cleanup-Image /RestoreHealth
+    Write-Completed
 
-# Windows Search purge and re-initialization
-Write-Header -header "Purging Windows Search..."
-net stop WSearch
-Remove-Item /S /Q "C:\ProgramData\Microsoft\Search"
-Remove-Item -Path "HKCU:Software\Microsoft\Windows Search" -Recurse
-Remove-ItemProperty -Path "HKLM:Software\Microsoft\Windows Search" -Name "SetupCompletedSuccessfully" -WhatIf
-Write-Completed
+    # WinSxS cleanup ResetBase
+    Write-Header -header "Cleaning up WinSxS..."
+    DISM /Online /Cleanup-Image /StartComponentCleanup /ResetBase
+    Write-Completed
 
-# Update Drivers
+    # Windows Search purge and re-initialization
+    Write-Header -header "Purging Windows Search..."
+    net stop WSearch
+    Remove-Item /S /Q "C:\ProgramData\Microsoft\Search"
+    Remove-Item -Path "HKCU:Software\Microsoft\Windows Search" -Recurse
+    Remove-ItemProperty -Path "HKLM:Software\Microsoft\Windows Search" -Name "SetupCompletedSuccessfully" -WhatIf
+    Write-Completed
 
-# Defragmentation
-# TODO: Refactor Defragmentation so can be read from config
-MyDefrag.exe -r DataDiskWeekly.MyD -v G: -v F: -v J: -v H:
+    # Reset run status
+    $config | ForEach-Object {$_.NeedsReboot=0}
+    $config | ConvertTo-Json -depth 32| set-content -Path config.json
+
+    # TODO: Update Drivers
+
+    # Defragmentation
+    if ($config.DataDrives.length -gt 0)
+    {
+        $DataDrives = $config.DataDrives
+        $DataDrives = $DataDrives -split " "
+    
+        for ($i = 0; $i -lt $DataDrives.length; $i++)
+        {
+            $DataDrives[$i] = "-v " + $DataDrives[$i]
+        }
+    
+        $Defrag = $DataDrives -join ' '
+    
+        MyDefrag.exe -r DataDiskWeekly.MyD $Defrag
+    }
+}
